fix(OUT-1679) : added RFC3339 date parsing for createdDate and overall folder structure refactoring.
