+++ src/app/api/tasks/templates/public/[id]/route.ts
@@ -1,4 +1,4 @@
 import { withErrorHandler } from '@/app/api/core/utils/withErrorHandler'
-import { getTaskTemplatePublic } from '@/app/api/tasks/templates/templates.controller'
+import { getTaskTemplatePublic } from '@/app/api/tasks/templates/public/public.controller'
 
 export const GET = withErrorHandler(getTaskTemplatePublic)
