+++ src/app/api/tasks/templates/public/route.ts
@@ -1,4 +1,4 @@
 import { withErrorHandler } from '@/app/api/core/utils/withErrorHandler'
-import { getTaskTemplatesPublic } from '@/app/api/tasks/templates/templates.controller'
+import { getTaskTemplatesPublic } from '@/app/api/tasks/templates/public/public.controller'
 
 export const GET = withErrorHandler(getTaskTemplatesPublic)
